+++ app/src/main/java/tumble/app/tumble/presentation/views/bookmarks/Calendar/BookmarkCalendarView.kt
@@ -1,30 +1,21 @@
 package tumble.app.tumble.presentation.views.bookmarks.Calendar
 
-import android.content.ClipData.Item
 import android.icu.util.Calendar
 import android.os.Build
-import android.util.Log
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.background
-import androidx.compose.foundation.gestures.Orientation
-import androidx.compose.foundation.gestures.scrollable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.pager.VerticalPager
 import androidx.compose.foundation.pager.rememberPagerState
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -32,27 +23,20 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.hilt.navigation.compose.hiltViewModel
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.cancel
-import kotlinx.coroutines.coroutineScope
-import kotlinx.coroutines.launch
-import tumble.app.tumble.extensions.presentation.toColor
-import tumble.app.tumble.presentation.components.buttons.CompactEventButtonLabel
 import tumble.app.tumble.presentation.viewmodels.BookmarksViewModel
 import tumble.app.tumble.presentation.views.general.CustomProgressIndicator
 import tumble.app.tumble.utils.month_date
 import kotlin.math.abs
-
 
 @RequiresApi(Build.VERSION_CODES.O)
 @OptIn(ExperimentalFoundationApi::class, ExperimentalCoroutinesApi::class)
@@ -125,7 +109,8 @@ fun BookmarkCalendarView(
             }
         }
         item{
-        BottomSheet() }
+            BottomSheet()
+        }
         item{
             Spacer(modifier = Modifier.height(60.dp))
         }
@@ -164,41 +149,6 @@ fun MonthName(
     )
 }
 
-@RequiresApi(Build.VERSION_CODES.O)
-@Composable
-fun BottomSheet(
-    viewModel: BookmarksViewModel = hiltViewModel(),
-) {
-    viewModel.selectedDate?.let {
-        val events = viewModel.bookmarkData.calendarEventsByDate.getOrDefault(it, emptyList())
-        if (events.isEmpty()) {
-            Row(
-                modifier = Modifier.fillMaxSize(),
-                verticalAlignment = Alignment.Top
-            ) {
-                Text(
-                    text = "No events for this date",
-                    color = MaterialTheme.colors.onBackground,
-                    fontSize = 20.sp
-                )
-            }
-        } else {
-            events.forEach { event ->
-                Row(modifier = Modifier
-                    .padding(bottom = 5.dp)
-                    .height(120.dp)) {
-                    CompactEventButtonLabel(
-                        event = event,
-                        color = event.course?.color?.toColor() ?: Color.Red
-                    )
-                }
-            }
-        }
-    }
-}
-
-
-
 @Composable
 fun DaysOfTheWeek(){
     Row(
@@ -219,4 +169,8 @@ fun DaysOfTheWeek(){
             }
         }
     }
+}
+
+enum class DaysOfWeek(){
+    Mon,Tur,Wed,Thu,Fri,Sat,Sun
 }
